make consistent use of 'GOOG' as openTypeOS2VendorID
Prior to this commit the `openTypeOS2VendorID` was not consistent across all
font variants. In particular, the italic variants used `pyrs` as vendor ID,
whereas the rest of the variants used `GOOG`. This appears to confuse certain
applications (such as Emacs), which failed to recognize the italic variants of
the font.